Revert "Render close button inside Header instead of injecting a node"
This reverts commit 8385973dd3df0dff95ef7437bfd1e1852d2b3cd6.
