+++ packages/react/src/modal/modal.tsx
@@ -1,3 +1,4 @@
+import { Close } from '@obosbbl/grunnmuren-icons-react';
 import { cx } from 'cva';
 import { useContext } from 'react';
 import { ButtonContext } from 'react-aria-components/Button';
@@ -15,7 +16,10 @@ import {
 } from 'react-aria-components/Modal';
 import { DEFAULT_SLOT, Provider } from 'react-aria-components/slots';
 
+import { Button } from '../button';
 import { HeaderContext } from '../content';
+import { translations } from '../translations';
+import { useLocale } from '../use-locale';
 
 type DialogTriggerProps = RACDialogTriggerProps;
 
@@ -66,14 +70,26 @@ const Modal = ({
   fullscreen = false,
   ...restProps
 }: ModalProps) => {
+  const locale = useLocale();
   return (
     <Provider
       values={[
         [
           HeaderContext,
           {
-            // Header renders the dismiss button itself; we only pass the close handler.
-            _onClose: isDismissable ? () => onOpenChange?.(false) : undefined,
+            // The close button is injected into the Header. RAC Dialog supports
+            // one "close" slot out of the box, which we utilize here.
+            _action: isDismissable ? (
+              <Button
+                slot="close"
+                variant="tertiary"
+                className="data-focus-visible:outline-focus-inset px-2.5!"
+                aria-label={translations.close[locale]}
+                onPress={() => onOpenChange?.(false)}
+              >
+                <Close />
+              </Button>
+            ) : undefined,
           },
         ],
       ]}
